Disclaimer
This note is just supportive material for object Oriented Programing in C++.This note may not cover all aspect of learning Object Oriented Programming in C++, nor these be taken as primary source of information. Students are encouraged to follow the core text books and reference book prescribed by the university.
For theory portion of this note various text books as well as freely available material from internet sources were consulted .Contents in this note are copyrighted to the instructor and authors of original texts where applicable.
Reproduction, distribution, republication, and or re-transmission of this note are prohibited without Prior Permission.
The content is this note is made to error free as far as possible. If you found any error in this note please directly report to pradippaudel89@gmail.com
I always expect your valuable suggestion, feedback as well criticism (if any) for the betterment of this note.

©2022, Pradip Paudel